feat(lib.luakittens): add luaKITTENS parsing
The luaKITTENS library is an attempt to redefine a type system in lua.
It consists of a parser for parsing luaKITTEN type expressions and a
validator which can then take a type and check if it complies with a
given kitten.

Currently only the parsing part is done and it requires some more
polishing as well.

The parser is implemented in lpeg, but I will rewrite it to a more
sophisticated solution later, which will allow for much better error
diagnostics.