feat(lib.luakittens): remove duplicates from alternative types
The grammar allows for alternative types, e.g. `string|number|(fn,
bool)`. This syntax, however, allows for specifying the same type twice.

I haven't found a way to disable that in the syntax itself, so the
resulting parse is postprocessed in the end to not carry additional,
unnecessary types around.